Create coroutine_stack header
